♻️ rename "rest" to "routes" to fit marketplace-core style
